Move some methods to the abstract class
